% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{authorization}
\alias{authorization}
\title{AWS request Authorization using Signature version 4 method.
Your credentials should be a list object with names of 'access' and 'secret'.
This function is not intended to be called directly; it is called by awsRestRequest}
\usage{
authorization(credentials, httpMethod = "GET", region = "us-east-1",
  service = "s3", host = "amazonaws.com", bucket = NULL, path = NULL,
  queryParameters = NULL, content = NULL, headers = NULL)
}
\description{
AWS signature 4:
http://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html
}
