% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credsFromInstanceMetadata.R
\name{credsFromInstanceMetadata}
\alias{credsFromInstanceMetadata}
\title{Uses curl to pull EC2 instance metadata which contains access, secret,
   and token for the Instance Role... not using a role, shame on you.}
\usage{
credsFromInstanceMetadata()
}
\description{
Uses curl to pull EC2 instance metadata which contains access, secret,
   and token for the Instance Role... not using a role, shame on you.
}
