% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCredentialFilePath.R
\name{genCredentialFilePath}
\alias{genCredentialFilePath}
\title{genCredentialFilePath}
\usage{
genCredentialFilePath(prefix = NULL, credentialFileName = "credentials",
  configFileName = "config")
}
\arguments{
\item{prefix}{(string) the path to where credential/config files are stored}

\item{credentialFileName}{(string) the arg name pretty much says it}

\item{configFileName}{(string) the arg name pretty much says it}
}
\description{
Identify the full path to source credential/config files
}
\details{
If no prefix is provided, attempts to identify the default path given the OS detacted.  
The default parameters should be sufficient unless you have customized the credential/config files.
It is assumed your credentials file follows the guidelines at:
[AWS CLI Configuration and Credential Files'](https://docs.aws.amazon.com/cli/latest/userguide/cli-config-files.html)
}
\examples{
genCredentialFilePath(
    prefix = '/path/to/files',
    credentialFileName = 'specialNameCredentials',
    configFileName = 'specialNameConfig'
)

}
\seealso{
Other credential management functions: \code{\link{AWSRootCredentials}},
  \code{\link{AWSTemporaryCredentials}},
  \code{\link{RcurlAWS}},
  \code{\link{credsFromInstanceMetadata}},
  \code{\link{getCredentials}}, \code{\link{getProfiles}},
  \code{\link{getSTSCredentials}},
  \code{\link{parseCredentialFile}},
  \code{\link{stsGenContentString}},
  \code{\link{tempCredentialHandler}}
}
\concept{credential management functions}
