% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSRootCredentials.R
\name{getCredentials}
\alias{getCredentials}
\title{getCredentials}
\usage{
getCredentials(AWS_ACCESS_KEY_ID = NULL, AWS_SECRET_ACCESS_KEY = NULL,
  profileName = NULL, credentialPrefix = NULL, credentialFileName = NULL,
  configFileName = NULL)
}
\arguments{
\item{AWS_ACCESS_KEY_ID}{Your AWS root account access id}

\item{AWS_SECRET_ACCESS_KEY}{Your AWS root account secret key}

\item{profileName}{Name of the profile you wish to invoke
'defualt' is used unless otherwise explicitly declared}

\item{sourceFile}{Location of credentials file ex. ~/.aws/.credential}

\item{.silent}{Doesn't do anything but is reserved to warning suppression or verbosity setting}
}
\description{
Worker function to find credentials in multiple search locations.
}
\details{
The initializer function of AWSRootCredentials object.
}
