% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCredentialFile.R
\name{parseCredentialFile}
\alias{parseCredentialFile}
\title{This isn't a super robust file parser, but will work well if AWS credential|config file guidelines are followed.
read lines of the credential or config file then iterate over forming a list of profiles.}
\usage{
parseCredentialFile(filePath = NULL)
}
\description{
This isn't a super robust file parser, but will work well if AWS credential|config file guidelines are followed.
read lines of the credential or config file then iterate over forming a list of profiles.
}
