% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcurlAWS.R
\docType{package}
\name{RcurlAWS}
\alias{RcurlAWS}
\alias{RcurlAWS-package}
\title{RcurlAWS: a package for R intended to facilitate use of AWS services without external dependencies.}
\description{
Primary development is around credential management (both root and STS), somewhat robust use of cerdential/config files, and generic API requests.  Aws Root and Temporary credential management implemented using R6 classes.  STS is supported to generate temporary credentials.
}
\details{
Helper functions to access AWS services via package::curl

\itemize{
  \item limited functionality is implemented in wrappers which obscure the worker functions 
  \item In theroy any HTTP(S) style request to the AWS REST API is possible with awsRestRequest 
  \item NEED TO add service-level [S3, SQS, etc.] modules 
  \item NEED TO ensuring all settings in a config file are used in service calls 
}
}
\seealso{
Other credential management functions: \code{\link{AWSRootCredentials}},
  \code{\link{AWSTemporaryCredentials}},
  \code{\link{credsFromInstanceMetadata}},
  \code{\link{genCredentialFilePath}},
  \code{\link{getCredentials}}, \code{\link{getProfiles}},
  \code{\link{getSTSCredentials}},
  \code{\link{parseCredentialFile}},
  \code{\link{stsGenContentString}},
  \code{\link{tempCredentialHandler}}

Other service wrappers: \code{\link{awsRestRequest}},
  \code{\link{getSTSCredentials}},
  \code{\link{stsGenContentString}}
}
\concept{credential management functions}
\concept{service wrappers}
