% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCredentialFile.R
\name{parseCredentialFile}
\alias{parseCredentialFile}
\title{This isn't a super robust file parser, but will work well if AWS credential|config file guidelines are followed.
read lines of the credential or config file then iterate over forming a list of profiles.}
\usage{
parseCredentialFile(filePath = NULL)
}
\description{
This isn't a super robust file parser, but will work well if AWS credential|config file guidelines are followed.
read lines of the credential or config file then iterate over forming a list of profiles.
}
\seealso{
Other credential management functions: \code{\link{AWSRootCredentials}},
  \code{\link{AWSTemporaryCredentials}},
  \code{\link{RcurlAWS}},
  \code{\link{credsFromInstanceMetadata}},
  \code{\link{genCredentialFilePath}},
  \code{\link{getCredentials}}, \code{\link{getProfiles}},
  \code{\link{getSTSCredentials}},
  \code{\link{stsGenContentString}},
  \code{\link{tempCredentialHandler}}
}
\concept{credential management functions}
