% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSTemporaryCredentials.R
\name{tempCredentialHandler}
\alias{tempCredentialHandler}
\title{tempCredentialHandler}
\usage{
tempCredentialHandler(rootCredentials = NULL, roleArn = NULL,
  RoleSessionName = NULL, MFADeviceSerialNumber = NULL, Duration = "3600")
}
\arguments{
\item{rootCredentials}{(AWSRootCredentials) Your root credential object to feed STS calls}

\item{roleArn}{(string) the ARN of the role you want to assume}

\item{RoleSessionName}{(string) An identifier for the request; this can be used for user auditing on AWS.}

\item{MFADeviceSerialNumber}{(string) The id number (serial | arn) of your MFA device.}

\item{Duration}{(string) Duration, in seconds, of the role session; controls time-to-expire of temporary credentials.}
}
\description{
Initializes a call to STS or pulls instance meta data
}
\details{
This function isn't intended to be called directly; it is a method of the AWSTemporaryCredentials class.
Initializes a call to STS or pulls instance meta data.
   Exactly which depends on the arguments you pass.
1. If root credentials are passed, it will call STS.
   - Only the root credentials and roleArn are required.
   - If you have an account with MFA, you must pass your MFA SN.
   - This only works with Assume Role for now, thus the requirement of a roleARN.

2. Otherwise, it will try to curl instance metadata and get temporary credentials from there.

For additional information on STS requests refer to [STS Requests](https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_request.html)
}
